+++ javascript/react/src/components/StepEditor/StepEditor.styles.ts
