+++ javascript/react/src/components/StepEditor/index.ts
@@ -2,5 +2,5 @@
  * Copyright (c) Microsoft. All rights reserved.
  * Licensed under the MIT license. See LICENSE file in the project.
  */
-export * from './StepListItem.js'
-export * from './StepListItem.types.js'
+export * from './StepEditor.js'
+export * from './StepEditor.types.js'
