+++ javascript/react/src/components/TableCommands/StepEditorModal.types.ts
@@ -4,8 +4,8 @@
  */
 import type { IModalProps } from '@fluentui/react'
 
-import type { StepListItemProps } from '../StepListItem/index.js'
+import type { StepEditorProps } from '../StepEditor/index.js'
 
-export interface StepEditorModalProps extends IModalProps, StepListItemProps {
+export interface StepEditorModalProps extends IModalProps, StepEditorProps {
 	target?: string
 }
